pyproject.toml
[project]
name = "ciscou-2385-radkit-mcp"
version = "1.0.0"
description = "Cisco RADKit MCP Server - CISCOU Demo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cisco-radkit-client==1.9.0",
    "cisco-radkit-common==1.9.0",
    "cisco-radkit-service==1.9.0",
    "fastmcp==2.13.1",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "questionary>=2.0.0"
]

[project.scripts]
ciscou-radkit-mcp = "ciscou_radkit_mcp.radkit_mcp_server:main"
ciscou-radkit-onboarding = "ciscou_radkit_mcp.radkit_onboarding:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ciscou_radkit_mcp"]